+++ spec/lumberjack/sidekiq_spec.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+require "spec_helper"
+
+RSpec.describe Lumberjack::Sidekiq do
+  describe "VERSION" do
+    it "has a version number" do
+      expect(Lumberjack::Sidekiq::VERSION).not_to be nil
+    end
+  end
+end